% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{alleledb_beta_binomial}
\alias{alleledb_beta_binomial}
\title{optimize beta-binomial parameters}
\usage{
alleledb_beta_binomial(counts, minN = 6, binSize = 40)
}
\description{
optimize beta-binomial parameters
}
