% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asb}
\alias{asb}
\title{SNVs associated with allele-specific binding (ASB)}
\format{A data frame with 8806 rows and 12 variables}
\source{
@source \url{http://alleledb.gersteinlab.org/download/}
}
\usage{
asb
}
\description{
SNVs associated with allele-specific binding (ASB)
}
\details{
\code{
  ===========
  CHANGELOG
  ===========

  v2.1
  --same results from v2.0, with additional columns for alternate alleles and p values from the betabinomial test


  v2.0 
  --results produced from updated scripts (available under 'Scripts' section, v2.0, http://alleledb.gersteinlab.org/docs/)
  --minor debugs in scripts


  v1.0
  --results in publication

  ---------------------------------------------------

  The columns are defined as follows:
  1:chr		 : chromosome
  2:start 	 : SNV start position (0-based)
  3:end 		 : SNV end position (1-based)
  4:TF_indiv_ASB   : (in ASB) transcription factor ; individual ID (as in the 1000 Genomes Project) ; ASB
    TF_indiv_accB  : transcription factor ; individual ID (as in the 1000 Genomes Project) ; accessible SNVs related to non-ASB
    gene_indiv_ASE : (in ASE) HGNC gene symbol or NA if not in any gene ; individual ID ; ASE
    gene_indiv_accE: HGNC gene symbol or NA if not in any gene ; individual ID ; accessible SNVs related to non-ASE
  5:ref_allele	 : reference allele
  6:alt_allele	 : alternate allele (multiple alleles are separated by ';')
  7:cA		 : read counts for A
  8:cC		 : read counts for C
  9:cG		 : read counts for G
  10:cT		 : read counts for T
  11:p.betabinom   : p values calculated from the betabinomial test in AlleleDB pipeline

  ** gene regions are defined in GENCODE v17.
}
}
\keyword{datasets}
