% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbalance.R
\name{bisect}
\alias{bisect}
\title{bisect}
\usage{
bisect(p, p_sim, p_choice, fdr, fdr_threshold, by, distrib = "binomial",
  b = 0, w, p_thresh)
}
\arguments{
\item{p}{p}

\item{p_sim}{p_sim}

\item{p_choice}{p_choice}

\item{fdr}{fdr}

\item{fdr_threshold}{fdr_threshold}

\item{by}{by}

\item{distrib}{distribution (binomial or betabinomial)}

\item{b}{overdispersion parameter}

\item{w}{w}

\item{p_thresh}{p_thresh}
}
\value{
something
}
\description{
bisect
}
