% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nulldistrib.R
\name{change_counts_into_density}
\alias{change_counts_into_density}
\title{Change "counts" into density}
\usage{
change_counts_into_density(d.combined.sorted.binned)
}
\arguments{
\item{d.combined.sorted.binned}{The combined, sorted, and binned
pseudodensity}
}
\value{
The combined, sorted, and binned density
}
\description{
Change "counts" into density
}
\seealso{
\code{\link{bin_according_to_empirical_distribution}}
}
