% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{chenimbalance_color_palette}
\alias{chenimbalance_color_palette}
\title{Provide a color palette}
\usage{
chenimbalance_color_palette()
}
\value{
A character vector representing the color palette
}
\description{
Provide a color palette
}
