% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_overdispersion_parameter.R
\name{choose_overdispersion_parameter}
\alias{choose_overdispersion_parameter}
\title{Choose overdispersion parameter}
\usage{
choose_overdispersion_parameter(w_grad, w, empirical, sse, minN = 6,
  binSize = 40, b_choice = 0, r_sta = 0, r_end = 0.99,
  r_by = 0.1)
}
\arguments{
\item{data}{A data frame with columns "total" and "allelicRatio"}

\item{bins}{Breakpoints for bins}
}
\value{
Histogram of allelic ratios
}
\description{
Minimize sse for betabinomials
}
