% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{choose_probability_of_success_parameter}
\alias{choose_probability_of_success_parameter}
\title{Choose probability of success parameter}
\usage{
choose_probability_of_success_parameter(w_grad, w, empirical, sse, b,
  minN = 6, binSize = 40, prob_choice = 0.1, r_sta = 0.2,
  r_end = 0.99, r_by = 0.1, n_cores = detectCores())
}
\arguments{
\item{w_grad}{w_grad}

\item{w}{w}

\item{empirical}{empirical}

\item{sse}{sse}

\item{b}{overdispersion parameter}

\item{minN}{minimum coverage level}

\item{binSize}{approximate number of bins}

\item{r_sta, r_end}{start and end of the parameter range}

\item{r_by}{step of the parameter range}

\item{n_cores}{number of cores to use}
}
\value{
list
}
\description{
Minimize sse for betabinomials
}
