% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{color_palette}
\alias{color_palette}
\title{Provide a color palette}
\usage{
color_palette()
}
\value{
A character vector representing the color palette
}
\description{
Provide a color palette
}
\seealso{
\code{\link{choose_overdispersion_parameter}}
}
