% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_overdispersion_parameter.R
\name{compute_betabinom_null}
\alias{compute_betabinom_null}
\title{Compute a beta-binomial null distribution}
\usage{
compute_betabinom_null(k)
}
\arguments{
\item{k}{The overdispersion parameter}
}
\value{
the beta-binomial null distribution
}
\description{
Compute a beta-binomial null distribution
}
