% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleledb_calcOverdispersion.R
\name{empirical_allelic_ratio}
\alias{empirical_allelic_ratio}
\title{Empirical allelic Ratio}
\usage{
empirical_allelic_ratio(data, bins, maxN, minN = 6)
}
\arguments{
\item{data}{A data frame with columns "total" and "allelicRatio"}

\item{bins}{Breakpoints for bins}
}
\value{
Histogram of allelic ratios
}
\description{
Empirical allelic Ratio
}
\details{
Plot right closed interval left open (range] for x axis
note that you have pseudozeroes as counts in your data so thats fine
}
