% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbalance.R
\name{fp}
\alias{fp}
\title{FP}
\usage{
fp(w, p, p_thresh, distrib = "binomial", b = 0)
}
\arguments{
\item{w}{w}

\item{p}{Numeric. The binomial parameter.}

\item{p_thresh}{p_thresh}
}
\value{
something
}
\description{
FP
}
