% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graded_weights.R
\name{graded_weights_for_sse_calculation}
\alias{graded_weights_for_sse_calculation}
\title{Graded weights for SSE calculation}
\usage{
graded_weights_for_sse_calculation(r_min, r_max, bins)
}
\arguments{
\item{r_min}{Range minimum}

\item{r_max}{Range maximum}

\item{bins}{Breakpoints mapped to unit interval}
}
\value{
Graded weights for SSE calculation
}
\description{
Graded weights for SSE calculation
}
