% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion.R
\name{optimize_overdispersion_parameter}
\alias{optimize_overdispersion_parameter}
\title{Optimize the overdispersion parameter}
\usage{
optimize_overdispersion_parameter(w_grad, w, b_and_sse, b_choice, sse,
  empirical, counter, minN = 6, p = 0.5, binSize = 40, r_by = 0.1,
  n_cores = detectCores())
}
\arguments{
\item{w_grad}{w_grad}

\item{w}{w}

\item{b_and_sse}{b_and_sse}

\item{b_choice}{b_choice}

\item{empirical}{empirical}

\item{counter}{counter}

\item{minN}{minimum coverage level}

\item{p}{binomial probability of success parameter}

\item{binSize}{approximate number of bins}

\item{r_by}{r_by}

\item{n_cores}{number of cores to use}
}
\value{
list
}
\description{
Minimize sse for betabinomials
}
