% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{optimize_probability_of_success_parameter}
\alias{optimize_probability_of_success_parameter}
\title{Optimize the probability of success parameter}
\usage{
optimize_probability_of_success_parameter(w_grad, prob_and_sse,
  prob_choice, sse, empirical, counter, b, minN = 6, binSize = 40,
  r_by = 0.1, n_cores = detectCores())
}
\arguments{
\item{w_grad}{w_grad}

\item{prob_and_sse}{b_and_sse}

\item{prob_choice}{b_choice}

\item{empirical}{empirical}

\item{counter}{counter}

\item{b}{overdispersion parameter}

\item{minN}{minimum coverage level}

\item{binSize}{approximate number of bins}

\item{r_by}{r_by}

\item{n_cores}{number of cores to use}
}
\value{
list
}
\description{
Minimize sse for betabinomials
}
