% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleledb_calcOverdispersion.R
\name{weight_by_empirical_counts}
\alias{weight_by_empirical_counts}
\title{Weight by empirical counts}
\usage{
weight_by_empirical_counts(total)
}
\arguments{
\item{total}{Vector containing coverage per variant}
}
\value{
A one-column matrix of weights for each coverage level
}
\description{
Weight by empirical counts
}
\seealso{
\code{\link{weighted_expected_binomial}}
}
