% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nulldistrib.R
\name{weight_with_actual_counts_in_empirical}
\alias{weight_with_actual_counts_in_empirical}
\title{Weight each coverage level with actual counts in empirical}
\usage{
weight_with_actual_counts_in_empirical(d.combined, d, w, i, k, ptr, minN = 6)
}
\arguments{
\item{d.combined}{The combined pseudodistribution}

\item{d}{The parametric density value}

\item{w}{w}

\item{k}{k}

\item{minN}{Integer. The minimum coverage level.}
}
\value{
The combined pseudodistribution weighted by the empirical
  distribution
}
\description{
Weight each coverage level with actual counts in empirical
}
\seealso{
\code{\link{nulldistrib}}
}
