% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_expected_binomial.R
\name{weighted_expected_binomial}
\alias{weighted_expected_binomial}
\title{Weighted expected binomial}
\usage{
weighted_expected_binomial(w, minN = 6, p = 0.5, binSize = 40,
  distrib = "binomial", b = 0)
}
\arguments{
\item{binSize}{Approximate number of bins}

\item{total}{Vector containing coverage per variant}
}
\value{
weighted expected binomial values
}
\description{
Plot the binomial distribution for each n
}
\details{
dbinom(seq(0,500),n=500,p=0.5) gives the pdf of 0-500, at n=500, p=0.5 in
binomial distrib

weight each probability by the number of SNPs with n reads
}
